First Commit
Initial commit of the CourseCommerceManager contract. This contract allows for the management of a commerce system for courses. In this system, the owner of the contract can add new products, each of which represents a course. The courses are identified by an ID, the address of the owner, the name of the course, and a price. Users can buy a course, which creates a new sale record that tracks the buyer's address and the date of the sale. The total number of sales is also tracked. The contract currently does not support transferring funds or limiting purchases to available courses, which will be added in future commits