Update #2
Now the buyProduct function only stores the product id instead of the entire product. The direct transfer of payment to the creator's address has been removed, now the ethers stay within the smart contract's address.

A ownerWithdraw function has been created which allows for withdrawal if it's the owner requesting it and if there are enough funds to do it.

A function has been created to obtain the details of a product based on its ID with related checks.

A function has been created to get the details of a sales order based on its ID plus checks, it also calls the function for product details.

The file has been reformatted to be more readable and descriptions have been added.